Update installer script and create new dist script for windows. todo: find out hwo to invoke PyInstaller from python
